Create Game.py - controller of PacDude game
This commit makes the Game Class because it will act as the 
controller of the PacDude Game. The class contains an initializer
that has the variables required to make the game work.

It will be implemented using PyGame because it is a 
sustainable environment for writing simple games. 
When run, the player will be able to experience
a real-time game against a computer, where a 
score CGPA requirement has to be met in order to win. 

This Game class is important because it incorporates all the 
various components needed all in one class in order to 
make the game work.